+++ CHANGELOG.md
@@ -11,7 +11,9 @@
 
 # 0.2.5-SNAPSHOT (2025-09-14)
 
-* upload answers
+- upload answers
+* tasks/div-answers do not wrap.
+
 
 # 0.2.4 (2025-09-14)